+++ test/functional/repository/find-options-locking/find-options-locking.ts
@@ -465,6 +465,7 @@ describe("repository > find options > locking", () => {
                         entityManager
                             .getRepository(Post)
                             .findOne({
+                                where: { id: 1 },
                                 lock: { mode: "pessimistic_write", tables: [] },
                             })
                             .should.be.rejectedWith(
@@ -491,6 +492,7 @@ describe("repository > find options > locking", () => {
                         entityManager
                             .getRepository(Post)
                             .findOne({
+                                where: { id: 1 },
                                 relations: ["author"],
                                 lock: {
                                     mode: "pessimistic_write",
@@ -512,6 +514,7 @@ describe("repository > find options > locking", () => {
                     return connection.manager.transaction((entityManager) => {
                         return Promise.all([
                             entityManager.getRepository(Post).findOne({
+                                where: { id: 1 },
                                 relations: ["author"],
                                 lock: {
                                     mode: "pessimistic_write",
@@ -519,6 +522,7 @@ describe("repository > find options > locking", () => {
                                 },
                             }),
                             entityManager.getRepository(Post).findOne({
+                                where: { id: 1 },
                                 relations: ["author"],
                                 lock: { mode: "pessimistic_write" },
                             }),
@@ -530,6 +534,7 @@ describe("repository > find options > locking", () => {
                     return connection.manager.transaction((entityManager) => {
                         return Promise.all([
                             entityManager.getRepository(Post).findOne({
+                                where: { id: 1 },
                                 relations: ["author"],
                                 lock: {
                                     mode: "pessimistic_write",
@@ -539,6 +544,7 @@ describe("repository > find options > locking", () => {
                             entityManager
                                 .getRepository(Post)
                                 .findOne({
+                                    where: { id: 1 },
                                     relations: ["author"],
                                     lock: { mode: "pessimistic_write" },
                                 })
@@ -561,6 +567,7 @@ describe("repository > find options > locking", () => {
                 return connection.manager.transaction((entityManager) => {
                     return Promise.all([
                         entityManager.getRepository(Post).findOne({
+                            where: { id: 1 },
                             relations: ["author"],
                             lock: {
                                 mode: "pessimistic_read",
@@ -568,6 +575,7 @@ describe("repository > find options > locking", () => {
                             },
                         }),
                         entityManager.getRepository(Post).findOne({
+                            where: { id: 1 },
                             relations: ["author"],
                             lock: {
                                 mode: "pessimistic_write",
@@ -575,6 +583,7 @@ describe("repository > find options > locking", () => {
                             },
                         }),
                         entityManager.getRepository(Post).findOne({
+                            where: { id: 1 },
                             relations: ["author"],
                             lock: {
                                 mode: "pessimistic_partial_write",
@@ -582,6 +591,7 @@ describe("repository > find options > locking", () => {
                             },
                         }),
                         entityManager.getRepository(Post).findOne({
+                            where: { id: 1 },
                             relations: ["author"],
                             lock: {
                                 mode: "pessimistic_write_or_fail",
@@ -589,6 +599,7 @@ describe("repository > find options > locking", () => {
                             },
                         }),
                         entityManager.getRepository(Post).findOne({
+                            where: { id: 1 },
                             relations: ["author"],
                             lock: {
                                 mode: "for_no_key_update",
@@ -614,6 +625,7 @@ describe("repository > find options > locking", () => {
                 return connection.manager.transaction((entityManager) => {
                     return Promise.all([
                         entityManager.getRepository(Post).findOne({
+                            where: { id: 1 },
                             join: {
                                 alias: "post",
                                 innerJoinAndSelect: {